Refactors PublishedContentType
PublishedContentType was previously using a custom PublishedSystemProperties class, because it differs from regular SystemProperties because of the naming of revision vs version. This caused unnecessary duplication, and it also broke the structure where resources implement ResourceInterface.
